refactor(alerter): extract anomaly_queries.go
Move anomaly candidate, embedding, baseline, and acknowledgment
functions. Extract scanAcknowledgedAnomalyAlert helper.

Part of #75.
